requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
passlib[bcrypt]
bcrypt
python-jose[cryptography]
cryptography
pandas
numpy
langchain
langchain-core
langchain-ollama
apscheduler
pydantic
python-dotenv
requests
jwt